update to address comments
