+++ src/RaceDetect.cpp
@@ -98,7 +98,7 @@ Report race::detectRaces(llvm::Module *module, DetectRaceConfig config) {
       }
 
       // No race if guaranteed to be executed by same thread
-      if (ompAnalysis.inSameGuardedTID(write, other)) return;
+      if (ompAnalysis.guardedBySameTID(write, other)) return;
 
       // Lastprivate code will only be executed by one thread
       // Model lastprivate by assuming lastprivate code cannot race with other last private code
